# Fixed some issues with type stability