--- a/src/linalg/LinAlg.jl
+++ b/src/linalg/LinAlg.jl
@@ -54,9 +54,9 @@ function eigs(A; kwargs...)
     schr =partialschur(A; kwargs...)
     vals, vectors = partialeigen(schr[1])
     reved = (kwargs[:which] == LR() || kwargs[:which] == LM())
-    k = get(kwargs, :nev, length(vals))
+    k::Int = get(kwargs, :nev, length(vals))
     k = min(k, length(vals))
-    perm = 1:k
+    perm = collect(1:k)
     if vals[1] isa(Real)
         perm = sortperm(vals, rev=reved)
         perm = perm[1:k]