--- a/src/linalg/nonbacktracking.jl
+++ b/src/linalg/nonbacktracking.jl
@@ -6,7 +6,7 @@ Return a non-backtracking matrix `B` and an edgemap storing the oriented
 edges' positions in `B`.
 
 Given two arcs ``A_{i j}` and `A_{k l}` in `g`, the
-non-backtraking matrix ``B`` is defined as
+non-backtracking matrix ``B`` is defined as
 
 ``B_{A_{i j}, A_{k l}} = δ_{j k} * (1 - δ_{i l})``
 """