--- a/src/shortestpaths/bellman-ford.jl
+++ b/src/shortestpaths/bellman-ford.jl
@@ -115,7 +115,7 @@ function enumerate_paths(state::AbstractPathState, vs::Vector{T}) where T<:Integ
     all_paths = Vector{Vector{T}}(undef, num_vs)
     for i = 1:num_vs
         all_paths[i] = Vector{T}()
-        index = vs[i]
+        index::T = vs[i]
         if parents[index] != 0 || parents[index] == index
             while parents[index] != 0
                 push!(all_paths[i], index)
@@ -125,7 +125,7 @@ function enumerate_paths(state::AbstractPathState, vs::Vector{T}) where T<:Integ
             reverse!(all_paths[i])
         end
     end
-    all_paths
+    return all_paths
 end
 
 enumerate_paths(state::AbstractPathState, v) = enumerate_paths(state, [v])[1]