--- a/src/shortestpaths/bellman-ford.jl
+++ b/src/shortestpaths/bellman-ford.jl
@@ -108,14 +108,15 @@ will return a vector (indexed by destination vertex) of paths from source `v`
 to all other vertices. In addition, `enumerate_paths(state, v, d)` will return
 a vector representing the path from vertex `v` to vertex `d`.
 """
-function enumerate_paths(state::AbstractPathState, vs::Vector{T}) where T<:Integer
+function enumerate_paths(state::AbstractPathState, vs::Vector{<:Integer})
     parents = state.parents
+    T = eltype(parents)
 
     num_vs = length(vs)
     all_paths = Vector{Vector{T}}(undef, num_vs)
     for i = 1:num_vs
         all_paths[i] = Vector{T}()
-        index::T = vs[i]
+        index = T(vs[i])
         if parents[index] != 0 || parents[index] == index
             while parents[index] != 0
                 push!(all_paths[i], index)