This PR should fix some type instability issues. It does not fix all of them and in some cases does only reduce some of the issues in a functions but not all of them. Especially working with `Channel` seems to be tricky to do in a completely type stable way.
I've checked by hand using different types, so there is also the possibility that I have overlooked something.
 
For those that do not know it, the method `@descend_code_warntype` of the package [Cthulhu.jl](https://github.com/vchuravy/Cthulhu.jl) was very helpful.