@sbromberger 2019-01-28
Do these changes actually improve timings? I wasn't able to see any speedup in `barabasi_albert!()`, for example, but I confess I didn't take a long look at it.