@simonschoelly 2019-01-28
To be honest, I did not do any benchmarks. The speedup might be more visible when using other types than the standard `Int` and  `Float64`, as in these cases, we often have type instability.

Here are some benchmarks for `barabasi_albert`. On a side node, I realised, that `barabasi_albert` takes the graph `eltype` from its second argument and not from the first one, as it probably should.

```
#old 

julia> @btime barabasi_albert(Int16(2^14), 3)
  4.968 ms (42298 allocations: 4.04 MiB)
{16384, 49143} undirected simple Int64 graph

#new

julia> @btime barabasi_albert(Int16(2^14), 3)
  4.807 ms (42297 allocations: 4.04 MiB)
{16384, 49143} undirected simple Int64 graph

# ==============================================

#old

julia> @btime barabasi_albert(Int16(2^14), Int16(3))
  4.705 ms (42298 allocations: 2.80 MiB)
{16384, 49143} undirected simple Int16 graph

#new

julia> @btime barabasi_albert(Int16(2^14), Int16(3))
  4.623 ms (42288 allocations: 2.80 MiB)
{16384, 49143} undirected simple Int16 graph
```