@simonschoelly 2019-01-28
Here are some more benchmarks. Some functions such as `simplecyclecount` did not even work before for `eltype != Int`.

```
# biconnected_components

## old

julia> g = squash(loadsnap(:facebook_combined));
julia> @btime biconnected_components(g);
  676.243 μs (9126 allocations: 414.42 KiB)

## new 

julia> g = squash(loadsnap(:facebook_combined));
julia> @btime biconnected_components(g);
  604.298 μs (222 allocations: 175.03 KiB)

# label_propagation

## old

julia> g = squash(loadsnap(:facebook_combined));
julia> @btime label_propagation(g);
  4.222 ms (14 allocations: 56.78 KiB)

## new

julia> g = squash(loadsnap(:facebook_combined));
julia> @btime label_propagation(g);
  4.245 ms (14 allocations: 56.78 KiB)

# cycle_basis

## old

julia> g = squash(loadsnap(:facebook_combined));
julia> @btime cycle_basis(g);
  362.402 ms (8042967 allocations: 134.73 MiB)

## new

julia> g = squash(loadsnap(:facebook_combined));
julia> @btime cycle_basis(g);
  19.159 ms (201568 allocations: 13.37 MiB)
```