@simonschoelly 2019-02-23
Would it be possible to get some code review here? I think we should merge this PR better sooner than later (if everything is ok with it) as it affects a lot of functions, so merging later might lead to conflicts.