@sbromberger 2019-02-24
PS: those `cycle_basis` performance improvements astound me. Why did we get such a dramatic speedup?