@simonschoelly 2019-02-24
One of the reasons might have been
```
if root == nothing
         root = pop!(gnodes)
end
```
which means, that root has to be of type `Union{Nothing, eltype(g)`.  This union type would then be carried into the mainloop. I replaced `root` by `r` in the mainloop and ensured that `r` is not an union type.

The second thing that could have been a reason is this:
```
used = Dict(root => [])
```
`[]` has eltype `Any`, so this is not good for perfomance.